App is yellow and a few working features at the moment.
First would have to be the edit function in the saved screen
Second is the cross hair on the camera screen. (the cross hair is just for show atm)

Wanting to make the image crop down to the size of the cross hair.

About to start working on a setting screen then follow that up by working on out putting multiple predictions and their percentages.